Changed how test class invokes the TestDataProvider class.
The Civi\Test helper functions seem to prevent the callback from firing in
some cases, which means $this->data may be undefined for the second test in a class.
Also, the database seems to get wiped (or reverted to some pre-test state) between
tests, so it's better to just run this code before each test.

For consistency with other DB-interacting tests, restored use of
TransactionalInterface, which per documentation is responsible for rolling
back DB changes between tests.